backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-openai
langchain-google-genai
langchain-groq
python-dotenv
sse-starlette
pydantic
pydantic-settings
httpx
